商城业务-秒杀服务-秒杀效果完成

+++ mall-seckill/src/main/java/com/grain/mall/seckill/to/SeckillSkuRedisTo.java
@@ -35,7 +35,7 @@ public class SeckillSkuRedisTo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
      * 每人限购数量
      */
